deleted calendar container and refactored calender to render from Moods
