--- src/containers/Calendar/Calendar.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import CalendarView from './CalendarView';
-
-const Calendar = () => {
-  return <CalendarView />;
-};
-
-export default Calendar;
